[core] -Fix: create.c integer ID count is long int to match object field's type
git-svn-id: svn://repo.hu/pcb-rnd/trunk@2938 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/create.c
@@ -53,7 +53,7 @@
 /* ---------------------------------------------------------------------------
  * some local identifiers
  */
-static int ID = 1;							/* current object ID; incremented after */
+static long int ID = 1;							/* current object ID; incremented after */
 				/* each creation of an object */
 
 static bool be_lenient = false;
